+++ src/anz_rosetta_csv/JsonTableSchema/JsonTableSchema.py
@@ -17,7 +17,12 @@
 
 import json
 import sys
-import csvdatatypes
+
+try:
+   import csvdatatypes
+except ModuleNotFoundError:
+   from src.anz_rosetta_csv.JsonTableSchema import csvdatatypes
+
 
 class FormatError(Exception): pass
 class DuplicateFieldName(Exception): pass
@@ -27,7 +32,7 @@ class JSONTableSchema(object):
 
    __format_version__ = "1.0-pre3.1-partial-implementation"
 
-   required_field_descriptor_keys = ["name"] 
+   required_field_descriptor_keys = ["name"]
    optional_field_descriptor_keys_strings = ["title", "type", "description"]
    optional_field_descriptor_keys_lists = ["constraints"]
    optional_field_descriptor_keys_patterns = ["format"]
@@ -35,72 +40,72 @@ class JSONTableSchema(object):
 
    def __init__(self, json_string=None):
       # Initialise JSONTableSchema object, optionally from a JSON string
-      
+
       self.fields = []
       self.format_version = self.__format_version__
-      
+
       if json_string is not None:
          try:
             self.read_json(json.loads(json_string))
          except ValueError:
             sys.stderr.write("Invalid JSON object the likely cause. Please chack and try again.")
-        
+
    def read_json(self, json_string):
-      
+
       if "fields" not in json_string:
          raise FormatError("JSON array `fields' must be present JSON Table Schema hash.")
- 
+
       field_list = json_string["fields"]
       if not isinstance(field_list, list):
          raise FormatError("JSON key `fields' must be array")
 
       for i, stanza in enumerate(field_list):
-          
+
          if not isinstance(stanza, dict):
             err_str = "Field descriptor %d must be a dictionary" % idx
             raise FormatError(err_str)
-         
-         for key in self.required_field_descriptor_keys: 
+
+         for key in self.required_field_descriptor_keys:
             if key not in stanza:
                err_tmpl = "Field descriptor %d must contain key `%s'" % (i, key)
                raise FormatError(err_tmpl)
-         
+
          for key in stanza:
             if key not in self.required_field_descriptor_keys and key not in self.optional_field_descriptor_keys_strings \
                and key not in self.optional_field_descriptor_keys_lists and key not in self.optional_field_descriptor_keys_patterns:
                err_tmpl = "Field descriptor %d shouldn't contain key `%s'" % (i, key)
                raise FormatError(err_tmpl)
-            
+
          self.add_field(stanza)
          self.format_version = json_string.get("json_table_schema_version", self.__format_version__)
 
    @property
    def field_names(self):
       return [ i["name"] for i in self.fields ]
-            
+
    def add_field(self, field):
-      
+
       field_dict = {}
 
-      for key in self.required_field_descriptor_keys:      
+      for key in self.required_field_descriptor_keys:
          if not isinstance(field[key], (str, unicode)):
             raise FormatError("Field `name' must be a string")
          if field["name"] in self.field_names:
             raise DuplicateFieldName("field 'name'")
          field_dict[key] = field[key]
-      
+
       for key in self.optional_field_descriptor_keys_strings:
          if key in field:
             if not isinstance(field[key], (str, unicode)):
                raise FormatError("Field `name' must be a string")
-            
+
             if key == "type":
                self.check_type(field["type"], field["name"])
-               
+
             field_dict[key] = field[key]
-      
+
       #TODO: Complex types, format and constraints
-    
+
       self.fields.append(field_dict)
 
    def remove_field(self, field_name):
@@ -123,15 +128,14 @@ class JSONTableSchema(object):
       return csv_header[:-1]
 
    def check_type(self, field_type, field_name):
-  
+
       type_found = False
       for field_category in csvdatatypes.__valid_type_names__:
          for type in field_category:
             if field_type == type:
                type_found = True
                break
-      
+
       if type_found != True:
          err_tmpl = "Invalid type `%s' in field descriptor for `%s'" % (field_type, field_name)
          raise FormatError(err_tmpl)
-
