+++ src/anz_rosetta_csv/RosettaCSVGenerator.py
@@ -4,32 +4,40 @@
 import os
 import sys
 sys.path.append(r'JsonTableSchema/')
-import ConfigParser
-import JsonTableSchema
-from ProvenanceCSVHandlerClass import *
-from droidcsvhandlerclass import *
-from rosettacsvsectionsclass import RosettaCSVSections
-from ImportSheetGenerator import ImportSheetGenerator
+import configparser as ConfigParser
+
+try:
+   import JsonTableSchema
+except ModuleNotFoundError:
+   from src.anz_rosetta_csv.JsonTableSchema import JsonTableSchema
+
+try:
+    from ProvenanceCSVHandlerClass import *
+    from droidcsvhandlerclass import *
+    from rosettacsvsectionsclass import RosettaCSVSections
+    from ImportSheetGenerator import ImportSheetGenerator
+except ModuleNotFoundError:
+    pass
 
 class RosettaCSVGenerator:
 
    def __init__(self, droidcsv=False, exportsheet=False, rosettaschema=False, configfile=False, provenance=False):
-   
+
       if configfile is not False:
          self.config = ConfigParser.RawConfigParser()
-         self.config.read(configfile)   
-         
+         self.config.read(configfile)
+
          self.droidcsv = droidcsv
          self.exportsheet = exportsheet
-         
+
          #NOTE: A bit of a hack, compare with import schema work and refactor
          self.rosettaschema = rosettaschema
          self.readRosettaSchema()
-         
+
          #Grab Rosetta Sections
          rs = RosettaCSVSections(configfile)
          self.rosettasections = rs.sections
-         
+
          #set provenance flag and file
          self.prov = False
          if provenance is True:
@@ -38,7 +46,7 @@ class RosettaCSVGenerator:
             if self.config.has_option('provenance', 'file'):
                #Overwrite default, if manually specified...
                self.provfile = self.config.get('provenance', 'file')
-         
+
             self.provhash = "MD5"
             if self.config.has_option('application configuration', 'provhash'):
                self.provhash = self.config.get('application configuration', 'provhash')
@@ -46,10 +54,10 @@ class RosettaCSVGenerator:
          self.pathmask = self.__setpathmask__()
       else:
          sys.exit('No config file')
-         
+
       #Get some functions from ImportGenerator
       self.impgen = ImportSheetGenerator()
-      
+
       #List duplicate items to check...
       self.duplicateitemsaddedset = set()
 
@@ -65,11 +73,11 @@ class RosettaCSVGenerator:
       f = open(self.rosettaschema, 'rb')
       importschemajson = f.read()
       importschema = JsonTableSchema.JSONTableSchema(importschemajson)
-      
+
       importschemadict = importschema.as_dict()
       importschemaheader = importschema.as_csv_header()
 
-      self.rosettacsvheader = importschemaheader + "\n"  #TODO: Add newline in JSON Handler class? 
+      self.rosettacsvheader = importschemaheader + "\n"  #TODO: Add newline in JSON Handler class?
       self.rosettacsvdict = importschemadict['fields']
       f.close()
 
@@ -99,14 +107,14 @@ class RosettaCSVGenerator:
       else:
          #Fail but don't exit desirable(?) so as to see all errors at once
          sys.stderr.write("Filename comparison has failed. Check list control: " + listcontroltitle.encode('utf-8') + " vs. DROID export: " + droid_filename_title.encode('utf-8') + "\n")
-      
+
       return True
 
    #NOTE: itemtitle is title from Archway List Control...
    def grabdroidvalue(self, md5, itemtitle, subseries, field, rosettafield, pathmask):
-   
+
       #TODO: Potentially index droidlist by MD5 or SHA-256 in future...
-      returnfield = ""      
+      returnfield = ""
       for drow in self.droidlist:
          addtorow = False
 
@@ -119,23 +127,23 @@ class RosettaCSVGenerator:
             checksumfromdroid = drow['SHA256_HASH']
          else:
             sys.stderr.write("No hash available to use in DROID export.\n")
-            sys.exit(1)         
-         
+            sys.exit(1)
+
          if checksumfromdroid == md5:
             if self.compare_filenames_as_titles(drow, itemtitle):
 
                #Performance, only do more work, if we have to care about it...
                if checksumfromdroid in self.duplicates:
-               
-                  #recreate subseries so that we can do comparison for path alignment... 
+
+                  #recreate subseries so that we can do comparison for path alignment...
                   subseriesfromdroid = os.path.dirname(drow['FILE_PATH']).replace(self.subseriesmask, '')
-               
+
                   if subseries == subseriesfromdroid:
                      addtorow = True
                      self.duplicateitemsaddedset.add(subseries + "\\" + itemtitle + " checksum: " + checksumfromdroid)
                else:
                   addtorow = True
-                 
+
          if addtorow == True:
             droidfield = drow[rosettafield]
             if field == 'File Location':
@@ -144,10 +152,10 @@ class RosettaCSVGenerator:
                returnfield = os.path.dirname(droidfield).replace(pathmask, '').replace('\\','/') + "/"
             else:
                returnfield = droidfield
-               
+
       return returnfield
-     
-   def csvstringoutput(self, csvlist):   
+
+   def csvstringoutput(self, csvlist):
       #String output...
       csvrows = self.rosettacsvheader
 
@@ -156,19 +164,19 @@ class RosettaCSVGenerator:
       #Static ROW in CSV Ingest Sheet
       SIPROW = ['"",'] * len(self.rosettacsvdict)
       SIPROW[0] = '"SIP",'
-      
-      #SIP Title... 
+
+      #SIP Title...
       if self.config.has_option('rosetta mapping', 'SIP Title'):
          SIPROW[1] = '"' + self.config.get('rosetta mapping', 'SIP Title') + '",'
       else:
          SIPROW[1] = '"CSV Load",'
-     
+
       csvrows = csvrows + ''.join(SIPROW).rstrip(',') + '\n'
-      
+
       #write utf-8 BOM
-      # NOTE: Don't write UTF-8 BOM... Rosetta doesn't like this. 
+      # NOTE: Don't write UTF-8 BOM... Rosetta doesn't like this.
       #sys.stdout.write(u'\uFEFF'.encode('utf-8'))
-      
+
       for sectionrows in csvlist:
          rowdata = ""
          for sectionrow in sectionrows:
@@ -176,14 +184,14 @@ class RosettaCSVGenerator:
                rowdata = rowdata + fielddata + ','
             rowdata = rowdata.rstrip(',') + '\n'
          csvrows = csvrows + rowdata
-         
+
       #this is the best i can think of because ExLibris have named two fields with the same
       #title in CSV which doesn't help us when we're trying to use unique names for populating rows
       #replaces SIP Title with Title (DC)
-      csvrows = csvrows.replace('"Object Type","SIP Title"','"Object Type","Title (DC)"')         
-         
+      csvrows = csvrows.replace('"Object Type","SIP Title"','"Object Type","Title (DC)"')
+
       sys.stdout.write(csvrows)
-      
+
       for dupe in self.duplicateitemsaddedset:
          sys.stderr.write("Duplicates to monitor: " + dupe + "\n")
 
@@ -209,7 +217,7 @@ class RosettaCSVGenerator:
       return pathmask
 
    def populaterows(self, field, listcontrolitem, sectionrow, csvindex, rnumber):
-   
+
       #populate cell with static values from config file
       if self.config.has_option('static values', field):
          rosettafield = self.config.get('static values', field)
@@ -219,7 +227,7 @@ class RosettaCSVGenerator:
       if self.config.has_option('rosetta mapping', field):
          rosettafield = self.config.get('rosetta mapping', field)
          addvalue = listcontrolitem[rosettafield]
-         
+
          #****MULTIPLE ACCESS RESTRICTIONS****#
          #If the field we've got in the config file is Access, we need to
          #Then grab the Rosetta access code for the correct restriction status
@@ -231,15 +239,15 @@ class RosettaCSVGenerator:
 
          #place value into the cell within the row...
          sectionrow[csvindex] = self.add_csv_value(addvalue)
-         
+
       #if there is a mapping to a value in the droid export...
-      elif self.config.has_option('droid mapping', field):          
-         rosettafield = self.config.get('droid mapping', field)         
+      elif self.config.has_option('droid mapping', field):
+         rosettafield = self.config.get('droid mapping', field)
          ignorefield = self.handleprovenanceexceptions('ORIGINALNAME', sectionrow, field, csvindex, rnumber)
 
          #if ignorefield is still false, check our checksum field as well...
          if ignorefield == False:
-		    ignorefield = self.handleprovenanceexceptions('CHECKSUM', sectionrow, field, csvindex, rnumber)
+            ignorefield = self.handleprovenanceexceptions('CHECKSUM', sectionrow, field, csvindex, rnumber)
 
          if ignorefield == False:
             sectionrow[csvindex] = self.add_csv_value(self.grabdroidvalue(listcontrolitem['Missing Comment'], listcontrolitem['Title'], listcontrolitem['Sub-Series'], field, rosettafield, self.pathmask))
@@ -250,20 +258,20 @@ class RosettaCSVGenerator:
                if field == 'Event Identifier Type':
                   sectionrow[csvindex] = self.add_csv_value("EXTERNAL")
                if field == 'Event Identifier Value':
-                  sectionrow[csvindex] = self.add_csv_value("EXT_1")  
+                  sectionrow[csvindex] = self.add_csv_value("EXT_1")
                if field == 'Event Type':
-                  sectionrow[csvindex] = self.add_csv_value("CREATION")  
+                  sectionrow[csvindex] = self.add_csv_value("CREATION")
                if field == 'Event Description':
                   sectionrow[csvindex] = self.add_csv_value("Provenance Note")
                if field == 'Event Date':
-                  sectionrow[csvindex] = self.add_csv_value(p['NOTEDATE'])  
+                  sectionrow[csvindex] = self.add_csv_value(p['NOTEDATE'])
                if field == 'Event Outcome1':
                   sectionrow[csvindex] = self.add_csv_value("SUCCESS")
                if field == 'Event Outcome Detail1':
                   sectionrow[csvindex] = self.add_csv_value(p['NOTETEXT'])
-  
+
    def createrosettacsv(self):
-      
+
       self.subseriesmask = ''
       if self.duplicates:
          if self.config.has_option('path values', 'subseriesmask'):
@@ -271,53 +279,53 @@ class RosettaCSVGenerator:
          else:
             sys.stderr.write("We have duplicate checksums, ensure they don't align with duplicate filenames")
             sys.stderr.write("Warning: '[path values] subseriesmask' not set in configuration.")
-            
+
       CSVINDEXSTARTPOS = 2
       csvindex = CSVINDEXSTARTPOS
-      
+
       self.rnumber = 0
       fields = []
 
       for item in self.exportlist:
-      
+
          itemfixity = False
          itemoriginalpath = False
-      
+
          itemrow = []
-         
+
          #self.rosettasections, list of dictionaries generated from CFG file...
          for sections in self.rosettasections:
-            #sections, individual dictionaries from CFG file... 
-            
+            #sections, individual dictionaries from CFG file...
+
             #section row is entire length of x-axis in spreadsheet from CSV JSON Config file...
             sectionrow = ['""'] * len(self.rosettacsvdict)
-            
+
             #Add key to the Y-axis of spreadsheet from dict...
             sectionrow[0] = self.add_csv_value(sections.keys()[0])
-            
+
             #driven by CFG file, not JSON, so field occurs in CFG file first...
             #e.g. IE, REPRESENTATION, FILE, then each field in each of those...
             for field in sections[sections.keys()[0]]:
 
                #store for record level handling like provenance
                if field == 'Archway Unique ID (Object Identifier)':
-                  self.rnumber = item['Item Code']            
-               
-               #if we have a matching field in the cfg, and json, populate it... 
+                  self.rnumber = item['Item Code']
+
+               #if we have a matching field in the cfg, and json, populate it...
                if field == self.rosettacsvdict[csvindex]['name']:
                   self.populaterows(field, item, sectionrow, csvindex, self.rnumber)
                else:
                   #we have a misalignment between cfg and json...
-                  #TODO: Output a more useful error message? 
+                  #TODO: Output a more useful error message?
                   sys.exit("CSV configuration and schema file do not match. Look for missing fields in either. Failed on: " + str(field) + " " + str(self.rosettacsvdict[csvindex]['name']))
-               
+
                #increment csvindex along the x-axis...
                csvindex+=1
-            
+
             itemrow.append(sectionrow)
          fields.append(itemrow)
          csvindex=CSVINDEXSTARTPOS
-      
+
       self.csvstringoutput(fields)
 
    #TODO: unit tests for this...
@@ -346,25 +354,25 @@ class RosettaCSVGenerator:
          csvhandler = genericCSVHandler()
          exportlist = csvhandler.csvaslist(self.exportsheet)
          return exportlist
-   
+
    def readDROIDCSV(self):
       if self.droidcsv != False:
          droidcsvhandler = droidCSVHandler()
          droidlist = droidcsvhandler.readDROIDCSV(self.droidcsv)
          droidlist = droidcsvhandler.removefolders(droidlist)
-         return droidcsvhandler.removecontainercontents(droidlist)        
-      
+         return droidcsvhandler.removecontainercontents(droidlist)
+
    def export2rosettacsv(self):
       if self.droidcsv != False and self.exportsheet != False:
          self.droidlist = self.readDROIDCSV()
          self.exportlist = self.readExportCSV()
          #self.readRosettaSchema()  #NOTE: Moved to constructor... TODO: Refactor
-         
+
          if self.prov is True:
             provhandler = provenanceCSVHandler()
             self.provlist = provhandler.readProvenanceCSV(self.provfile)
             if self.provlist is None:
                self.prov = False
-         
+
          self.duplicates = self.listduplicates()
          self.createrosettacsv()
